requirements.txt
Box2D==2.3.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
fonttools==4.30.0
glfw==2.5.1
gpytorch==1.6.0
gym==0.24.1
gym-notices==0.0.6
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.4.0
lockfile==0.12.2
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pycparser==2.21
pygame==2.1.0
pyglet==1.2.4
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
